fix: fix prepared statement cache error
When updating gorm from 1.21.7 to 1.23.10 kafka migrations
start to return an error:
Could not migrate: ERROR: cached plan must not change result type (SQLSTATE 0A000)
This error happens due to in the kafka migrations we perform database
queries and schema table modifications in an interleaved way and
prepared statements are cached. This leads scenarios where a specific
table is queried and thus some prepared statements cached, then some
table schemas are modified and then when queries are performed again against
those queries are outdated cached prepared statements which causes the
observed error.
This behavior is likely a bug in gorm related to prepared statement
caching due to only updating gorm minor version should not break changes.
To workaround this for now this commit updates code to disable prepared
statements during kafka migrations. This includes the kafka migrations when
integration tests are run too.
During serve command execution, for connectors migrations and for the
integration tests (excluding the migrations part of them) the prepared
statements are enabled being the behavior the same as before this change.